data type fix

+++ app/src/main/java/com/jithin/weatherapp/network/WeatherResponse.kt
@@ -34,7 +34,7 @@ data class WeatherData(
     @SerializedName("clouds") val clouds: Clouds,
     @SerializedName("wind") val wind: Wind,
     @SerializedName("visibility") val visibility: Int,
-    @SerializedName("pop") val pop: Int,
+    @SerializedName("pop") val pop: Double,
     @SerializedName("sys") val sys: Sys,
     @SerializedName("dt_txt") val dtTxt: String
 )
